fast/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pymysql==1.1.0
SQLAlchemy==2.0.28
langchain-openai==0.0.8
langchain_community==0.0.28
langchain-core==0.1.32
langchain==0.1.12
email-validator==2.1.1
python-multipart==0.0.9